Fix infinite haproxies
Ok, not really. Consul-template launched a new one every time it
attempted to reload haproxy. This was because the command to reload
wasn't executing $(cat path/to/pid). The solution was pulling this out
into its own script. Now, the old haproxy will terminate correctly. It
just needed to send the FINISH signal to the correct (former) haproxy
instance.